tanggle: Visualization of Phylogenetic Networks

Offers functions for plotting split (or implicit) networks (unrooted, undirected) and explicit networks (rooted, directed) with reticulations extending. 'ggtree' and using functions from 'ape' and 'phangorn'. It extends the 'ggtree' package [@Yu2017] to allow the visualization of phylogenetic networks using the 'ggplot2' syntax. It offers an alternative to the plot functions already available in 'ape' Paradis and Schliep (2019) doi:10.1093/bioinformatics/bty633 and 'phangorn' Schliep (2011) doi:10.1093/bioinformatics/btq706 .

See also

Author

Maintainer: Klaus Schliep klaus.schliep@gmail.com (ORCID)

Authors:

Marta Vidal-Garcia

Claudia Solis-Lemus solislemus@wisc.edu (ORCID)

Leann Biancani

Eren Ada

L. Francisco Henao Diaz

Joshua Justison

Other contributors:

Guangchuang Yu [contributor]